+++ generic/tclDisassemble.c
@@ -1268,7 +1268,7 @@ int
 Tcl_DisassembleObjCmd(
     ClientData clientData,	/* What type of operation. */
     Tcl_Interp *interp,		/* Current interpreter. */
-    int objc,			/* Number of arguments. */
+    size_t objc,			/* Number of arguments. */
     Tcl_Obj *const objv[])	/* Argument objects. */
 {
     static const char *const types[] = {
