+++ generic/tclIORTrans.c
@@ -497,7 +497,7 @@ int
 TclChanPushObjCmd(
     TCL_UNUSED(void *),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const *objv)
 {
     ReflectedTransform *rtPtr;	/* Instance data of the new (transform)
@@ -742,7 +742,7 @@ int
 TclChanPopObjCmd(
     TCL_UNUSED(void *),
     Tcl_Interp *interp,
-    int objc,
+    size_t objc,
     Tcl_Obj *const *objv)
 {
     /*
